refactor: send error from action code (#12)
